refactor: cjsToEsm to esm
